+++ pkgs/http/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Add `MockClient.pngResponse`, which makes it easier to fake image responses.
 * Added the ability to fetch the URL of the response through `BaseResponseWithUrl`.
+* Add the ability to get headers as a `Map<String, List<String>` to
+  `BaseResponse`.
 
 ## 1.1.2